Changed distance to type uint16_t so it can accomodate bigger values. Changed formula for calculating distance from time.

+++ dist_measure.h
@@ -31,7 +31,7 @@ uint16_t HCCalculateTime();
  * Calculate the distance to object
  * @return distance in cm
  */
-uint8_t HCCalculateDistance();
+uint16_t HCCalculateDistance();
 
 #endif	/* DIST_MEASURE_H */
